requirements.txt.txt
streamlit

numpy

matplotlib
Pillow